Added OBD Test
Testing extracting information from my cars OBD2 port using an ELM327 OBD2 to USB reader.

The goal of this was to see if you could tell the difference between two drivers based on their patterns of acceleration from a complete stop. This was planned for a submission to the Azure IOT hackathon on devpost, but was ultimately abandoned due to difficulty with the platform.

If you're using linux you'll need to add yourself to the dialout group to access /dev/USB0, or run this as root.